Move server/client stuff around and annotate it
