+++ src/main/java/com/convallyria/taleofkingdoms/TaleOfKingdomsAPI.java
@@ -4,7 +4,7 @@ import com.convallyria.taleofkingdoms.client.packet.ClientPacketHandler;
 import com.convallyria.taleofkingdoms.common.scheduler.Scheduler;
 import com.convallyria.taleofkingdoms.common.schematic.ClientSchematicHandler;
 import com.convallyria.taleofkingdoms.common.schematic.SchematicHandler;
-import com.convallyria.taleofkingdoms.common.schematic.ServerSchematicHandler;
+import com.convallyria.taleofkingdoms.server.schematic.ServerSchematicHandler;
 import com.convallyria.taleofkingdoms.common.world.ConquestInstanceStorage;
 import com.convallyria.taleofkingdoms.managers.IManager;
 import com.convallyria.taleofkingdoms.managers.SoundManager;
