+++ src/main/java/com/convallyria/taleofkingdoms/common/entity/guild/FoodShopEntity.java
@@ -5,7 +5,7 @@ import com.convallyria.taleofkingdoms.client.gui.entity.FoodShopScreen;
 import com.convallyria.taleofkingdoms.common.entity.TOKEntity;
 import com.convallyria.taleofkingdoms.common.shop.ShopItem;
 import com.convallyria.taleofkingdoms.common.shop.ShopParser;
-import com.convallyria.taleofkingdoms.common.world.ClientConquestInstance;
+import com.convallyria.taleofkingdoms.client.schematic.ClientConquestInstance;
 import com.google.common.collect.ImmutableList;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
