+++ src/main/java/com/convallyria/taleofkingdoms/server/commands/debug/TaleOfKingdomsInvokeCommand.java
@@ -3,7 +3,7 @@ package com.convallyria.taleofkingdoms.server.commands.debug;
 import com.convallyria.taleofkingdoms.TaleOfKingdoms;
 import com.convallyria.taleofkingdoms.client.translation.Translations;
 import com.convallyria.taleofkingdoms.common.world.ConquestInstance;
-import com.convallyria.taleofkingdoms.common.world.ServerConquestInstance;
+import com.convallyria.taleofkingdoms.server.world.ServerConquestInstance;
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.context.CommandContext;
