+++ src/main/java/com/convallyria/taleofkingdoms/server/schematic/ServerSchematicHandler.java
@@ -1,7 +1,12 @@
-package com.convallyria.taleofkingdoms.common.schematic;
+package com.convallyria.taleofkingdoms.server.schematic;
 
 import com.convallyria.taleofkingdoms.TaleOfKingdoms;
 import com.convallyria.taleofkingdoms.TaleOfKingdomsAPI;
+import com.convallyria.taleofkingdoms.common.schematic.Schematic;
+import com.convallyria.taleofkingdoms.common.schematic.SchematicHandler;
+import com.convallyria.taleofkingdoms.common.schematic.SchematicOptions;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.math.BlockBox;
 import net.minecraft.util.math.BlockPos;
@@ -9,6 +14,7 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
 
+@Environment(EnvType.SERVER)
 public final class ServerSchematicHandler extends SchematicHandler {
 
     @Override
